requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.4.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
BM25==1.0.0
boto3==1.34.117
botocore==1.34.117
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.53.0
free_proxy==1.1.1
frozenlist==1.4.1
fsspec==2024.5.0
google==3.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
gradio==4.36.1
gradio_client==1.0.1
graphviz==0.20.3
greenlet==3.0.3
groq==0.8.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jiter==0.4.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.15
langchain-anthropic==0.1.11
langchain-aws==0.1.3
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-groq==0.1.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.67
language_data==1.2.0
logger==1.4
lxml==5.2.2
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
minify_html==0.15.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pg8000==1.31.2
pillow==10.3.0
playwright==1.43.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.10
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
scramp==1.4.5
scrapegraphai==1.5.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
tenacity==8.3.0
thinc==8.2.4
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
